praktikum 13 dan 14

+++ routes/web.php
@@ -29,13 +29,13 @@ Route::get('/dashboard', function () {
 
 Route::middleware('auth')->group(function () {
     //Route untuk menampilkan student
-    Route::get('admin/student', [StudentController::class, 'index']);
+    Route::get('admin/student', [StudentController::class, 'index'])->middleware('admin');
 
     //Route untuk menampilkan form tambah student
-    Route::get('admin/student/create', [StudentController::class, 'create']);
+    Route::get('admin/student/create', [StudentController::class, 'create'])->middleware('admin');
 
     //Route untuk mengirim data student
-    Route::post('admin/student/store', [StudentController::class, 'store']);
+    Route::post('admin/student/store', [StudentController::class, 'store'])->middleware('admin');
 
     // Route untuk menampilkan halaman edit student
     Route::get('admin/student/edit/{id}', [StudentController::class, 'edit']);
@@ -62,20 +62,20 @@ Route::middleware('auth')->group(function () {
     Route::delete('admin/student/delete/{id}', [StudentController::class, 'destroy']);
     
 
-    //Route untuk menampilkan form tambah student
+    //Route untuk menampilkan form tambah course
     Route::get('admin/course/create', [CourseController::class, 'create']);
 
-    //Route untuk mengirim data student
-    Route::post('admin/course/store', [CourseController::class, 'store']);
+    //Route untuk mengirim data course
+    Route::post('admin/course/store', [CourseController::class, 'store'])->middleware('admin');
 
-    // Route untuk menampilkan halaman edit student
-    Route::get('admin/course/edit/{id}', [CourseController::class, 'edit']);
+    // Route untuk menampilkan halaman edit course
+    Route::get('admin/course/edit/{id}', [CourseController::class, 'edit'])->middleware('admin');
 
-    // Route untuk menyimpan hasil update student
-    Route::put('admin/course/update/{id}', [CourseController::class, 'update']);
+    // Route untuk menyimpan hasil update course
+    Route::put('admin/course/update/{id}', [CourseController::class, 'update'])->middleware('admin');
 
-    // Route untuk menghapus student
-    Route::delete('admin/course/delete/{id}', [CourseController::class, 'destroy']);
+    // Route untuk menghapus course
+    Route::delete('admin/course/delete/{id}', [CourseController::class, 'destroy'])->middleware('admin');
 
 
     //Route untuk menampilkan course
